+++ Cargo.toml
@@ -44,9 +44,9 @@ sha3 = "*"
 rand = "0.8"
 
 # Used for GPU proving.
-shivini = "=0.154.9"
-proof-compression = "=0.154.9"
-zksync-gpu-prover = "=0.154.9"
+shivini = "=0.154.10"
+proof-compression = "=0.154.10"
+zksync-gpu-prover = "=0.154.10"
 
 [profile.release.package."blake2_with_compression_verifier"]
 opt-level = 1
